feat: external module for create Card
